requirements.txt
streamlit>=1.32
pandas>=2.0
numpy>=1.24
yfinance>=0.2.40
plotly>=5.18
gspread>=6.0
google-auth>=2.28
